Move from_relbench utility into torch_geometric/contrib/utils/
