+++ torch_geometric/contrib/__init__.py
@@ -4,6 +4,7 @@ import torch_geometric.contrib.transforms  # noqa
 import torch_geometric.contrib.datasets  # noqa
 import torch_geometric.contrib.nn  # noqa
 import torch_geometric.contrib.explain  # noqa
+import torch_geometric.contrib.utils  # noqa
 
 warnings.warn(
     "'torch_geometric.contrib' contains experimental code and is subject to "
