+++ torch_geometric/contrib/utils/__init__.py
@@ -1,0 +1,3 @@
+from ._relbench import from_relbench
+
+__all__ = ['from_relbench']
